smart-recruit-api/requirements.txt
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.3
flask-marshmallow==1.4.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.5
SQLAlchemy==2.0.48
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
